10/DJ6/telegram_bot_project/requirements.txt
# Django и DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Telegram Bot
aiogram==3.2.0
aiohttp==3.9.1

# Дополнительные зависимости
python-dotenv==1.0.0